chore(reorg): move checkbox to new file
