+++ src/data-workspace/inputs/index.js
@@ -1,3 +1,4 @@
 export * from './file-inputs.js'
 export * from './long-text.js'
 export * from './inputs.js'
+export * from './true-only-checkbox.js'
